+++ CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGELOG
  * Add `BackedEnum` support with `#[Argument]` and `#[Option]` inputs in invokable commands
  * Allow Usages to be specified via `#[AsCommand]` attribute.
  * Allow passing invokable commands to `Symfony\Component\Console\Tester\CommandTester`
+ * Add `#[Input]` attribute to support DTOs in commands
 
 7.3
 ---